requirements.txt
ninja
numpy<2
pybind11 
transformers==4.56.2
diffusers==0.35.1
PEFT==0.17.1
cupy-cuda12x 
bpy
deepspeed
einops 
accelerate==1.10.1
nvidia-cublas-cu12
bitsandbytes
wandb
kornia
open3d
trimesh
omegaconf
iopath
python-box
kiui
opencv-python
imageio[ffmpeg]
sentencepiece
gpytoolbox
pillow 
timm
pymeshlab
rembg
scipy
onnxruntime 
xatlas
coscmd
git+https://github.com/NVlabs/nvdiffrast.git